calc speed into reducer
