+++ src/modules/Game/duck/types.ts
@@ -1,4 +1,5 @@
 export type GameStatus = "settings" | "game" | "stop";
+export type Speed = "pause" | "play" | "slow" | "fast";
 export interface GameState {
   status: GameStatus;
   speed: number;
